Move function types back into functions/src/types
Shared types in a parent directory <rootProject>/types
is causing tsc transpiler to create a different tree-structure
within transpiled functions/lib javascript files.
It causing firebase-tools to fail importing index.js and
may cause further issues
