+++ functions/src/api/v1/rooms/joinRoom.ts
@@ -7,8 +7,8 @@ import {
     TTL_ACCESS_TOKEN_PARTICIPANT_SECONDS,
 } from './service/createTwilioClientToken'
 import { createRoom } from './createRoom'
-import { UID } from '../../../../../types/uid'
-import { Room, RoomId } from '../../../../../types/Room'
+import { UID } from '../../../types/uid'
+import { Room, RoomId } from '../../../types/Room'
 
 /**
  * @swagger
