+++ functions/src/api/v1/rooms/service/createTwilioClientToken.ts
@@ -1,5 +1,5 @@
-import { RoomId } from '../../../../../../types/Room'
-import { UID } from '../../../../../../types/uid'
+import { RoomId } from '../../../../types/Room'
+import { UID } from '../../../../types/uid'
 import * as AccessToken from 'twilio/lib/jwt/AccessToken'
 import { getTwilioEnv } from '../../../../firebase/env'
