+++ functions/src/db/addRoom.ts
@@ -1,7 +1,7 @@
 import { COLLECTION_ROOMS, DEFAULT_ROOM_TYPE } from './constants'
-import { RoomId } from '../../../types/Room'
+import { RoomId } from '../types/Room'
 import { db, serverTimestamp } from '../firebase/firebase'
-import { UID } from '../../../types/uid'
+import { UID } from '../types/uid'
 
 export const addRoom = async (
     userId: UID,
