+++ functions/src/db/getUserToken.ts
@@ -1,7 +1,7 @@
-import { BadRequestError } from '../api/errors/HttpError'
 import { getUser } from './getUser'
-import { UID } from '../../../types/uid'
-import { JWTToken } from '../../../types/JWT'
+import { JWTToken } from '../types/JWT'
+import { UID } from '../types/uid'
+import { BadRequestError } from '../api/errors/HttpError'
 
 export const getUserToken = async (userId: UID): Promise<JWTToken | null> => {
     const userData = await getUser(userId)
