+++ functions/src/types/JWT.ts
@@ -1,0 +1,9 @@
+import { UID } from './uid'
+
+export interface JWTData {
+    uid: UID
+    iat: number
+}
+
+export type JWTKey = string
+export type JWTToken = string
