+++ functions/src/types/NewRoomResponse.ts
@@ -1,0 +1,6 @@
+import { RoomId, RoomSid } from './Room'
+
+export interface NewRoomResponse {
+    roomId: RoomId
+    roomSid: RoomSid
+}
